- Mike Gist added csWeakRefHashReversible. - Mike Gist made the engine lists hold weak references to objects instead of references and fixed the fallout. - Mike Gist fixed many ref counting problems, mostly increfs without a decref.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/collections@29209 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/cstool/unusedresourcehelper.cpp
@@ -35,7 +35,7 @@ namespace CS
         iMaterialList *matList = engine->GetMaterialList();
 
         iRegionList *regList = engine->GetRegions();
-        for(int i=0; i<matList->GetCount(); i++)
+        for(size_t i=0; i<matList->GetCount(); i++)
         {
           int regionCount = 0;
           csArray<iRegion*> regionList;
@@ -113,7 +113,7 @@ namespace CS
         iMeshFactoryList *factList = engine->GetMeshFactories();
 
         iRegionList *regList = engine->GetRegions();
-        for(int i=0; i<factList->GetCount(); i++)
+        for(size_t i=0; i<factList->GetCount(); i++)
         {
           int regionCount = 0;
           csArray<iRegion*> regionList;
